Added the 2 Necessary Functions
Added a function for getting an estimate for Beta. Also started the main function that will call and optimize the first function. So far that's all it does; still need to add all the other outputs.

Tested the results on a linear model I already had loaded up, and the coefficients looked spot on.